+++ src/components/gui_rect.c
@@ -3,7 +3,7 @@
 static GUIRect GUI_RECTS_ARENA[1 << 14];
 static size_t N_GUI_RECTS = 0;
 
-GUIRect* gui_rect_create(size_t width, size_t height, Vec4 color) {
+GUIRect* gui_rect_create(float width, float height, Vec4 color) {
     GUIRect* gui_rect = &GUI_RECTS_ARENA[N_GUI_RECTS++];
 
     gui_rect->width = width;
