+++ src/components/gui_rect.h
@@ -2,9 +2,9 @@
 #include "../la.h"
 
 typedef struct GUIRect {
-    size_t width;
-    size_t height;
+    float width;
+    float height;
     Vec4 color;
 } GUIRect;
 
-GUIRect* gui_rect_create(size_t width, size_t height, Vec4 color);
+GUIRect* gui_rect_create(float width, float height, Vec4 color);
